docs_api/requirements.txt
fastapi
uvicorn[standard]
openai
langchain
langchain-chroma
chromadb
requests
python-dotenv
minio
pydantic-settings
python-docx